pyproject.toml
[project]
name = "mini-memes-hehe"
version = "0.1.0"
requires-python = ">=3.10"

dependencies = [
    "torch==2.2.2",
    "transformers==4.40.2",
    "pillow==11.0.0",
    "numpy==1.26.4",
    "ultralytics>=8.3.231",
    "sentence-transformers>=3.1.1",
]


